requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.6
kiwisolver==1.1.0
numpy==1.17.2
Pillow==5.4.1
psycopg2-binary==2.8.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.18.4
six==1.12.0
sqlparse==0.3.0
torch==1.3.0
torchvision==0.4.1
urllib3==1.22
whitenoise==4.1.4
